Migrate `ownership` test to `wit-bindgen test`
Nice test to make use of many-runners-but-one-test.
