--- crates/test-rust-wasm/src/bin/borrowing-duplicate-if-necessary.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/ownership/borrowing-duplicate-if-necessary.rs");
-
-fn main() {}
