--- crates/test-rust-wasm/src/bin/borrowing.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/ownership/borrowing.rs");
-
-fn main() {}
