--- crates/test-rust-wasm/src/bin/owning.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/ownership/owning.rs");
-
-fn main() {}
