+++ tests/runtime-new/rust/ownership/test.wit
@@ -12,7 +12,7 @@ interface both-list-and-resource {
     list-and-resource: func(a: thing);
 }
 
-world ownership {
+world runner {
     import lists: interface {
         foo: func(a: list<list<string>>) -> list<list<string>>;
     }
@@ -36,6 +36,30 @@ world ownership {
     }
 
     import both-list-and-resource;
+}
 
-    export foo: func();
+world test {
+    export lists: interface {
+        foo: func(a: list<list<string>>) -> list<list<string>>;
+    }
+
+    export thing-in: interface {
+        record thing {
+            name: string,
+            value: list<string>
+        }
+
+        bar: func(a: thing);
+    }
+
+    export thing-in-and-out: interface {
+        record thing {
+            name: string,
+            value: list<string>
+        }
+
+        baz: func(a: thing) -> thing;
+    }
+
+    export both-list-and-resource;
 }
