+++ tests/runtime/main.rs
@@ -15,7 +15,6 @@ use wit_parser::{Resolve, WorldId, WorldItem};
 
 mod flavorful;
 mod options;
-mod ownership;
 mod resource_alias;
 mod resource_floats;
 mod resource_import_and_export;
